+++ lib/features/trips/driver_profile/presentation/bloc/driver_profile_event.dart
@@ -1,0 +1,17 @@
+part of 'driver_profile_bloc.dart';
+
+abstract class DriverProfileEvent extends Equatable {
+  const DriverProfileEvent();
+
+  @override
+  List<Object?> get props => [];
+}
+
+class DriverProfileLoadRequested extends DriverProfileEvent {
+  const DriverProfileLoadRequested(this.tripId);
+
+  final String tripId;
+
+  @override
+  List<Object?> get props => [tripId];
+}